Sparse Tests: split into 2 files, add tests for failing CSC
